tx-19.4.27-1 移除错误代码，新增错误代码字典

+++ CoreNetWorkingLib/Classes/TXNetWorking/TXNetModel.h
@@ -10,7 +10,7 @@
 
 @interface TXNetModel : TXModel
 /** code值:0成功 非0时取值为错误代码 */
-@property (nonatomic,assign)int code;
+@property (nonatomic,assign)NSInteger code;
 /** data为服务器返回给用户的有效数据,data为空时,该属性不出现 */
 @property (nonatomic,strong)id data;
 /** msg值为返回消息,非0时数据有效.告知请求者错误消息,此消息严禁用于用户交互 */
